Point to main website instead of repo. Refs #16.
The website is likely to become the main point of reference for all
things copilot, whereas the specific links to files in the repo could
break more often.

This change removes links to examples and other files in the repo and
includes just one link to the main website on github.io.