Disabled sorting of output of interpreter. Refs #19.
The sorting was an effect of using a Map for storing the results of the
interpreter. This has now been implemented using a list of tuples.

Using a Map ensures us that the order of triggers does not matter. However the
user of Copilot could define the triggers as impure, and therefore the order
would matter. We want to be as transparent on the order as possible, and
therefore the interpreter should match the order instead of sorting them.